# Backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/wallets": {
      "post": {
        "operationId": "WalletsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWalletDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "wallets"
        ]
      },
      "get": {
        "operationId": "WalletsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WalletDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "wallets"
        ]
      }
    },
    "/v1/wallets/{id}": {
      "get": {
        "operationId": "WalletsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "wallets"
        ]
      }
    },
    "/v1/wallets/{id}/transactions": {
      "get": {
        "operationId": "WalletsController_history",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletHistoryPageDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "wallets"
        ]
      }
    },
    "/v1/wallets/{id}/deposits": {
      "post": {
        "operationId": "WalletsController_deposit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Unique key (8-128 chars, [A-Za-z0-9_.:-]) per logical operation. Retries with the same key and payload replay the original response.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepositDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Simulated synchronous top-up (phase-1 sandbox rail; use POST /funding/topups for the async bank flow).",
        "tags": [
          "wallets"
        ]
      }
    },
    "/v1/wallets/{id}/withdrawals": {
      "post": {
        "operationId": "WalletsController_withdraw",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Unique key (8-128 chars, [A-Za-z0-9_.:-]) per logical operation. Retries with the same key and payload replay the original response.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WithdrawDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Synchronous sandbox withdrawal (phase 1; use POST /funding/withdrawals for the async bank flow).",
        "tags": [
          "wallets"
        ]
      }
    },
    "/v1/fx/rates": {
      "get": {
        "operationId": "FxController_rates_",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FxRatesTableDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "fx"
        ]
      }
    },
    "/v1/fx/quotes": {
      "post": {
        "operationId": "FxController_quote",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuoteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FxQuoteDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "fx"
        ]
      }
    },
    "/v1/fx/conversions": {
      "post": {
        "operationId": "FxController_convert",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Unique key (8-128 chars, [A-Za-z0-9_.:-]) per logical operation. Retries with the same key and payload replay the original response.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteConversionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversionResultDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "fx"
        ]
      }
    },
    "/v1/admin/wallets/{id}": {
      "get": {
        "operationId": "AdminWalletsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminWalletDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin / wallets"
        ]
      }
    },
    "/v1/admin/wallets/{id}/status": {
      "post": {
        "operationId": "AdminWalletsController_setStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetWalletStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin / wallets"
        ]
      }
    },
...
  },
  "info": {
    "title": "PayCore API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "Currency": {
        "type": "string",
        "enum": [
          "PKR",
          "AED",
          "USD"
        ],
        "description": "Currencies the user can receive in (active wallets)"
      },
...
      "MoneyDto": {
        "type": "object",
        "properties": {
          "currency": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Currency"
              }
            ]
          },
          "amount": {
            "type": "string",
            "description": "Decimal string in major units, e.g. \"1250.50\"",
            "example": "1250.50"
          },
          "amountMinor": {
            "type": "string",
            "description": "Integer minor units as a string, e.g. \"125050\"",
            "example": "125050"
          }
        },
        "required": [
          "currency",
          "amount",
          "amountMinor"
        ]
      },
...
      "EntryType": {
        "type": "string",
        "enum": [
          "DEPOSIT",
          "WITHDRAWAL",
          "TRANSFER",
          "FX_CONVERSION",
          "REVERSAL",
          "ADJUSTMENT",
          "PAYMENT",
          "REFUND",
          "TOPUP",
          "WITHDRAWAL_HOLD",
          "WITHDRAWAL_PAYOUT",
          "SETTLEMENT",
          "SETTLEMENT_PAYOUT"
        ]
      },
...
      "CreateWalletDto": {
        "type": "object",
        "properties": {
          "currency": {
            "type": "string",
            "enum": [
              "PKR",
              "AED",
              "USD"
            ]
          }
        },
        "required": [
          "currency"
        ]
      },
      "WalletStatus": {
        "type": "string",
        "enum": [
          "ACTIVE",
          "FROZEN",
          "CLOSED"
        ]
      },
      "WalletDto": {
        "type": "object",
        "properties": {
          "currency": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Currency"
              }
            ]
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/WalletStatus"
              }
            ]
          },
          "id": {
            "type": "string"
          },
          "balance": {
            "$ref": "#/components/schemas/MoneyDto"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "currency",
          "status",
          "id",
          "balance",
          "createdAt"
        ]
      },
      "WalletHistoryItemDto": {
        "type": "object",
        "properties": {
          "type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/EntryType"
              }
            ]
          },
          "direction": {
            "enum": [
              "IN",
              "OUT"
            ],
            "type": "string"
          },
          "currency": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Currency"
              }
            ]
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true,
            "nullable": true
          },
          "postingId": {
            "type": "string"
          },
          "transactionId": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "amount": {
            "type": "string"
          },
          "amountMinor": {
            "type": "string"
          },
          "balanceAfter": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "direction",
          "currency",
          "metadata",
          "postingId",
          "transactionId",
          "description",
          "amount",
          "amountMinor",
          "balanceAfter",
          "createdAt"
        ]
      },
      "WalletHistoryPageDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WalletHistoryItemDto"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "items",
          "nextCursor"
        ]
      },
      "DepositDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "string"
          }
        },
        "required": [
          "amount"
        ]
      },
      "TransactionLegDto": {
        "type": "object",
        "properties": {
          "direction": {
            "enum": [
              "IN",
              "OUT"
            ],
            "type": "string"
          },
          "currency": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Currency"
              }
            ]
          },
          "walletId": {
            "type": "string"
          },
          "amount": {
            "type": "string"
          },
          "amountMinor": {
            "type": "string"
          },
          "balanceAfter": {
            "type": "string"
          }
        },
        "required": [
          "direction",
          "currency",
          "walletId",
          "amount",
          "amountMinor",
          "balanceAfter"
        ]
      },
      "TransactionDto": {
        "type": "object",
        "properties": {
          "type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/EntryType"
              }
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "COMPLETED"
            ]
          },
          "reversalOfId": {
            "type": "string",
            "nullable": true
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true,
            "nullable": true
          },
          "legs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransactionLegDto"
            }
          },
          "id": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "status",
          "reversalOfId",
          "metadata",
          "legs",
          "id",
          "description",
          "createdAt"
        ]
      },
      "WithdrawDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "string"
          },
          "pin": {
            "type": "string"
          }
        },
        "required": [
          "amount",
          "pin"
        ]
      },
      "FxPairDto": {
        "type": "object",
        "properties": {
          "from": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Currency"
              }
            ]
          },
          "to": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Currency"
              }
            ]
          },
          "midRate": {
            "type": "string"
          },
          "customerRate": {
            "type": "string"
          }
        },
        "required": [
          "from",
          "to",
          "midRate",
          "customerRate"
        ]
      },
      "FxRatesTableDto": {
        "type": "object",
        "properties": {
          "pairs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FxPairDto"
            }
          },
          "spreadBps": {
            "type": "number"
          },
          "feeBps": {
            "type": "number"
          }
        },
        "required": [
          "pairs",
          "spreadBps",
          "feeBps"
        ]
      },
      "CreateQuoteDto": {
        "type": "object",
        "properties": {
          "fromCurrency": {
            "type": "string",
            "enum": [
              "PKR",
              "AED",
              "USD"
            ]
          },
          "toCurrency": {
            "type": "string",
            "enum": [
              "PKR",
              "AED",
              "USD"
            ]
          },
          "sellAmount": {
            "type": "string",
            "description": "Amount of `fromCurrency` to convert (fee is charged on top)."
          }
        },
        "required": [
          "fromCurrency",
          "toCurrency",
          "sellAmount"
        ]
      },
      "FxQuoteStatus": {
        "type": "string",
        "enum": [
          "OPEN",
          "EXECUTED",
          "EXPIRED"
        ]
      },
      "FxQuoteDto": {
        "type": "object",
        "properties": {
          "fromCurrency": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Currency"
              }
            ]
          },
          "toCurrency": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Currency"
              }
            ]
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/FxQuoteStatus"
              }
            ]
          },
          "journalEntryId": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string"
          },
          "sell": {
            "$ref": "#/components/schemas/MoneyDto"
          },
          "fee": {
            "$ref": "#/components/schemas/MoneyDto"
          },
          "totalDebit": {
            "$ref": "#/components/schemas/MoneyDto"
          },
          "buy": {
            "$ref": "#/components/schemas/MoneyDto"
          },
          "midRate": {
            "type": "string"
          },
          "customerRate": {
            "type": "string"
          },
          "spreadBps": {
            "type": "number"
          },
          "feeBps": {
            "type": "number"
          },
          "expiresAt": {
            "type": "string"
          }
        },
        "required": [
          "fromCurrency",
          "toCurrency",
          "status",
          "journalEntryId",
          "id",
          "sell",
          "fee",
          "totalDebit",
          "buy",
          "midRate",
          "customerRate",
          "spreadBps",
          "feeBps",
          "expiresAt"
        ]
      },
      "ExecuteConversionDto": {
        "type": "object",
        "properties": {
          "quoteId": {
            "type": "string",
            "format": "uuid"
          },
          "pin": {
            "type": "string"
          }
        },
        "required": [
          "quoteId",
          "pin"
        ]
      },
      "ConversionResultDto": {
        "type": "object",
        "properties": {
          "quote": {
            "$ref": "#/components/schemas/FxQuoteDto"
          },
          "transaction": {
            "$ref": "#/components/schemas/TransactionDto"
          }
        },
        "required": [
          "quote",
          "transaction"
        ]
      },
      "AdminWalletDto": {
        "type": "object",
        "properties": {
          "currency": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Currency"
              }
            ]
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/WalletStatus"
              }
            ]
          },
          "id": {
            "type": "string"
          },
          "balance": {
            "$ref": "#/components/schemas/MoneyDto"
          },
          "createdAt": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "ledgerAccountId": {
            "type": "string"
          }
        },
        "required": [
          "currency",
          "status",
          "id",
          "balance",
          "createdAt",
          "userId",
          "ledgerAccountId"
        ]
      },
      "SetWalletStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "FROZEN",
              "CLOSED"
            ]
          },
          "reason": {
            "type": "string",
            "minLength": 5,
            "maxLength": 500
          }
        },
        "required": [
          "status",
          "reason"
        ]
      },
...
    }
  }
}